+++ src/colruyt/pcrsejb/service/dl/team/DbTeamService.java
@@ -75,7 +75,7 @@ public class DbTeamService extends DbService implements TeamService {
     public void deleteElement(Team element) {
         try (Connection conn = this.createConnection()) {
             PreparedStatement preparedStatement = conn.prepareStatement(DELETE_ELEMENT);
-            preparedStatement.setInt(1, (int) element.getTeamID());
+            preparedStatement.setInt(1, element.getTeamID());
             ResultSet rs = preparedStatement.executeQuery();
         } catch (SQLException e1) {
             e1.printStackTrace();
